+++ examples/docker/video-rag/Dockerfile
@@ -1,12 +1,10 @@
 FROM python:3.11
 
 COPY app.py app.py
-COPY chat.py chat.py
 COPY .env .env
 COPY requirements.txt requirements.txt
+RUN pip install torch==2.1.1 torchvision==0.16.1 --index-url https://download.pytorch.org/whl/cpu
 COPY static/ static/
 RUN pip install -r requirements.txt
-RUN apt install ffmpeg
-
 
 ENTRYPOINT ["solara", "run", "app.py", "--host=0.0.0.0", "--port=80"]
