+++ examples/docker/video-rag/app.py
@@ -1,10 +1,7 @@
-from dotenv import load_dotenv
-import os 
+from pathlib import Path
+from dataclasses import dataclass
 
-import openai
 import solara
-import solara.lab
-from solara.components.file_drop import FileDrop
 
 from haystack import Pipeline
 from haystack.components.embedders import SentenceTransformersTextEmbedder
@@ -12,23 +9,22 @@ from haystack.components.builders.prompt_builder import PromptBuilder
 from haystack.components.generators import GPTGenerator
 from elasticsearch_haystack.embedding_retriever import ElasticsearchEmbeddingRetriever
 from elasticsearch_haystack.document_store import ElasticsearchDocumentStore
-
-from chat import chatbox_css, Message, ChatBox
+from dotenv import load_dotenv
+import os
 
 load_dotenv(".env")
 openaikey = os.getenv("OPENAI")
-qdrant = os.getenv("qdrant")
+elastic_search_cloud_id = os.getenv("elastic_search_cloud_id")
 elastic_search_host = os.getenv("elastic_search_host")
 elastic_username = os.getenv("elastic_username")
 elastic_password = os.getenv("elastic_password")
-
+#
 # Build RAG pipeline
 print("Initializing QA pipeline")
-######## Complete this section #############
 prompt_template = """\
 Use the following context to answer the user's question in a friendly manner. \
     If the context provided doesn't answer the question - \
-        please respond with: "I don't know".
+        please respond with: "There is no information in my knowledge base about this".
 
 ### CONTEXT
 {% for doc in documents %}
@@ -42,6 +38,7 @@ Use the following context to answer the user's question in a friendly manner. \
 #document_store = ElasticsearchDocumentStore(hosts= "http://localhost:9200/")
 document_store = ElasticsearchDocumentStore(hosts=elastic_search_host,
                                             basic_auth=(elastic_username, elastic_password))
+
 prompt_builder = PromptBuilder(prompt_template)
 ############################################
 query_embedder = SentenceTransformersTextEmbedder()
@@ -58,6 +55,12 @@ pipeline.connect("query_embedder.embedding", "retriever.query_embedding")
 pipeline.connect("retriever.documents", "prompt_builder.documents")
 pipeline.connect("prompt_builder", "llm")
 
+###########################################
+# Solara app
+
+class State:
+    input = solara.reactive("")
+
 css = """
     .main {
         width: 100%;
@@ -72,53 +75,61 @@ css = """
 }
 """
 
-openai.api_key = os.getenv("OPENAI")
+chatbox_css = """
+.message {
+    max-width: 450px;
+    width: 100%;
+}
 
-pipeline.draw("question_answer_pipeline.png")
+.user-message, .user-message > * {
+    background-color: #f0f0f0 !important;
+}
 
-class State:
-    video_data = solara.reactive(None)
-    video_upload_error = solara.reactive("")
-    existing_video_selection = solara.reactive("")
-    input = solara.reactive("")
+.assistant-message, .assistant-message > * {
+    background-color: #9ab2e9 !important;
+}
 
-    @staticmethod
-    def select_existing_video(video_name):
-        # Assuming video_name is the name or path of the pre-existing video
-        State.existing_video_selection.value = video_name
-        State.video_data.value = None
-        State.video_upload_error.value = ""
+.avatar {
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  border: 2px solid transparent;
+  overflow: hidden;
+  display: flex;
+}
 
-    @staticmethod
-    def load_from_file(file):
-        allowed_extensions = ['.mp4', '.avi', '.mov']
-        if not any(file["name"].lower().endswith(ext) for ext in allowed_extensions):
-            State.video_upload_error.value = "Only MP4, AVI, MOV files are supported"
-            return
-        try:
-            # Assuming file is a video file, process as needed
-            State.video_data.value = file
-            # Add any additional processing you need for the video file here
-        except Exception as e:
-            State.video_upload_error.value = str(e)
-
-        try:
-            # Process the video file
-            video_path = State.save_temp_video(file)
-            audio_path = State.convert_video_to_audio(video_path)
-            chunked_audio_paths = State.chunk_audio(audio_path, 180000)  # Chunk length in ms
-
-            # Run the indexing pipeline
-            State.index_audio_chunks(chunked_audio_paths)
-        except Exception as e:
-            State.video_upload_error.value = str(e)
+.avatar img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+"""
 
 
-    @staticmethod
-    def reset():
-        State.video_data.value = None
-        State.video_upload_error.value = ""
-    
+@dataclass
+class Message:
+    role: str
+    content: str
+
+
+def ChatBox(message: Message) -> None:
+    solara.Style(chatbox_css)
+
+    align = "start" if message.role == "assistant" else "end"
+    with solara.Column(align=align):
+        with solara.Card(classes=["message", f"{message.role}-message"]):
+            if message.content:
+                with solara.Card():
+                    solara.Markdown(message.content)
+            
+
+        # Image reference: https://www.flaticon.com/free-icons/bot;
+        #                  https://www.flaticon.com/free-icons/use
+
+        with solara.HBox(align_items="center"):
+            image_path = Path(f"static/{message.role}-logo.png")
+            solara.Image(str(image_path), classes=["avatar"])
+            solara.Text(message.role.capitalize())
 
 @solara.component
 def Chat() -> None:
@@ -134,13 +145,19 @@ def Chat() -> None:
         [
             Message(
                 role="assistant",
-                content=f"Welcome. Please post your queries!"
+                content=f"Welcome. Please post your queries! My knowledge base\
+                    has been curated on a small collection of videos from NASA.  \
+                    This collection of videos consist of short clips that talk \
+                    about the topics: Mars Perseverance Rover.\
+                    Sample questions: \n\nWhat is the Mars Perseverance Rover? \
+                        What is the Mars Perseverance Rover mission? \
+                        Tell me about the helicopter on Mars."
             )
         ]
     )
     input, set_input = solara.use_state("")
 
-    def ask_rag():
+    def ask_rag(pipeline):
         try:
             input_text = State.input.value
             _messages = messages + [Message(role="user", content=input_text)]
@@ -154,7 +171,7 @@ def Chat() -> None:
             set_messages(_messages + [Message(role="assistant", content=rag_response)])
 
         except Exception as e:
-            set_messages(_messages + [Message(role="assistant", content=str(e))])
+            set_messages(_messages + [Message(role="assistant", content=f"Cannot answer your current question. Please try again {e}")])
 
     with solara.VBox():
         for message in messages:
@@ -165,49 +182,26 @@ def Chat() -> None:
             solara.InputText(label="Query", value=State.input, continuous_update=False)
 
     if State.input.value:
-        ask_rag()
-
-
+        ask_rag(pipeline)
 
 @solara.component
 def Page():
 
-    video_upload_error = State.video_upload_error.value
-    existing_videos = ["video1.mp4", "video2.mov"]  # List of pre-existing videos
-
     with solara.AppBarTitle():
-        solara.Text("Video Upload and Analysis App")
+        solara.Text("Deepen your understanding of our video collection through a Q&A AI assistant")
 
     with solara.Card(title="About", elevation=6, style="background-color: #f5f5f5;"):
-        solara.Markdown("Upload and analyze video content using advanced AI tools.")
+        with solara.Row(justify="center"):
+            solara.Image(image="static/nasa-logo.svg", width="100")  # Adjust width and height as needed
 
-    with solara.Sidebar():
-        with solara.Card("Controls", margin=0, elevation=0):
-            with solara.Column():
-                FileDrop(
-                    on_file=State.load_from_file,
-                    on_total_progress=lambda *args: None,
-                    label="Drag a video file here",
-                )
-
-                for video in existing_videos:
-                    solara.Button(video, on_click=lambda v=video: State.select_existing_video(v))
-
-                if State.existing_video_selection.value:
-                    solara.Info(f"Selected video: {State.existing_video_selection.value}")
-
-                if State.video_data.value:
-                    solara.Info("Video is successfully uploaded")
-                    # Code to display the uploaded video goes here
-
-                if video_upload_error:
-                    solara.Error(f"Error uploading video: {video_upload_error}")
-
-
+        solara.Markdown("Ask questions about our curated database of video using advanced AI tools. \n \
+                        This database is curated from the following list of videos: \n \
+                        https://images.nasa.gov/search?q=nasa%20perseverance%20rover&page=1&media=video&yearStart=2023&yearEnd=2024")
+               
     solara.Style(css)
     with solara.VBox(classes=["main"]):
         solara.HTML(
-            tag="h3", style="margin: auto;", unsafe_innerHTML="Chat with my video"
+            tag="h3", style="margin: auto;", unsafe_innerHTML="Chat with the assistant to answer questions about the video topics"
         )
 
         Chat()
