+++ examples/docker/video-rag/images/question_answer_pipeline.png
